Initial commit
(stars "z3t0") will give the total number of stars for my
repositories. Including those that I have contirbuted to, even if I do
not own them. Although, this may not count repositories that I have
been since removed from, or those where I did not fork..? Ie. those
that I have a member permission on are probably the only counted.

Also, this only considers public repositories ... private repositories
either do not have a concept for stars and if they did it would not be
relevant for the concept that stars serve.